+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Scrollimation slides can omit `scroll_range` to grow to fit their content.
+- Scrollimation slides now accept a `font_scale` field.
+
 ### Changed
 
 ### Deprecated